+++ frontend/app/src/shared/components/menu/domain/get-menu.ts
